requirements.txt
google-generativeai
streamlit
gTTs
pygame
pandas
speechrecognition
pyaudio
tensorflow